+++ isitfit/tagsSuggestBasic.py
@@ -12,13 +12,34 @@ def dump_df_to_csv(df_dump, csv_prefix):
     with tempfile.NamedTemporaryFile(prefix=csv_prefix, suffix='.csv', delete=False) as fh:
       logger.info(colored("Dumping data into %s"%fh.name, "cyan"))
       df_dump.to_csv(fh.name, index=False)
-
-      logger.info(colored("Done","cyan"))
-      logger.info(colored("Consider `pip3 install visidata` and then `vd %s` for further filtering or exploration."%fh.name,"cyan"))
-      logger.info(colored("More details about visidata at http://visidata.org/","cyan"))
+      return fh.name
 
 
-class TagsListener:
+MAX_ROWS = 10
+MAX_COLS = 5
+MAX_STRING = 20
+
+def display_df(title, df, csv_fn, shape):
+    from tabulate import tabulate
+    logger.info("")
+    logger.info(title)
+    df_show = df.head(n=MAX_ROWS)
+    df_show = df_show.applymap(lambda c: (c[:MAX_STRING]+'...' if len(c)>=MAX_STRING else c) if type(c)==str else c)
+
+    logger.info(tabulate(df_show, headers='keys', tablefmt='psql', showindex=False))
+    if shape[0] <= MAX_ROWS and shape[1] <= MAX_COLS:
+      return
+
+    # https://pypi.org/project/termcolor/
+    from termcolor import colored
+
+    logger.info("...")
+    logger.info(colored("Consider `pip3 install visidata` and then `vd %s` for further filtering or exploration."%csv_fn,"cyan"))
+    logger.info(colored("More details about visidata at http://visidata.org/","cyan"))
+
+
+
+class TagsSuggestBasic:
 
   def __init__(self):
     # boto3 ec2 and cloudwatch data
@@ -27,7 +48,14 @@ class TagsListener:
     self.tags_list = []
     self.tags_df = None
 
-  def fetch(self, onlyNames):
+  def prepare(self):
+    pass
+
+  def tags_to_dict(self, ec2_obj):
+    tags_dict = {x['Key']: x['Value'] for x in ec2_obj.tags if x['Key']=='Name'}
+    return tags_dict
+
+  def fetch(self):
     logger.info("Counting EC2 instances")
     n_ec2_total = len(list(self.ec2_resource.instances.all()))
     logger.warning("Found a total of %i EC2 instances"%n_ec2_total)
@@ -42,17 +70,13 @@ class TagsListener:
       if ec2_obj.tags is None:
         tags_dict = {}
       else:
-        if onlyNames:
-          tags_dict = {x['Key']: x['Value'] for x in ec2_obj.tags if x['Key']=='Name'}
-        else:
-          tags_dict = {x['Key']: x['Value'] for x in ec2_obj.tags}
+        tags_dict = self.tags_to_dict(ec2_obj)
 
       tags_dict['instance_id'] = ec2_obj.instance_id
       self.tags_list.append(tags_dict)
 
     # convert to pandas dataframe when done
-    tags_df = self._list_to_df()
-    return tags_df
+    self.tags_df = self._list_to_df()
 
 
   def _list_to_df(self):
@@ -65,9 +89,18 @@ class TagsListener:
       return df
 
 
-  def suggest(self, names_df):
+  def suggest(self):
       logger.info("Generating suggested tags")
       from .tagsImplier import TagsImplierMain
-      tags_implier = TagsImplierMain(names_df)
-      suggested_df = tags_implier.imply()
-      return suggested_df
+      tags_implier = TagsImplierMain(self.tags_df)
+      self.suggested_df = tags_implier.imply()
+      self.csv_fn = dump_df_to_csv(self.suggested_df, 'isitfit-tags-suggest-')
+
+
+  def display(self):
+    display_df(
+      "Suggested tags:",
+      self.suggested_df,
+      self.csv_fn,
+      self.suggested_df.shape
+    )